backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib[bcrypt]
pyotp
pydantic[email]
python-dotenv
web3
redis
fastapi-limiter
greenlet
sentry-sdk
stripe
pydantic-settings
python-multipart
pytest
prometheus-fastapi-instrumentator
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
# Background worker dependencies
rq
rq-dashboard
schedule
# AI/ML dependencies
sentence-transformers
scikit-learn
numpy
pandas
user-agents
